added user authentication and ability to delete songs from library

+++ routes/user-library.js
@@ -9,6 +9,9 @@ import path from "path"
 export function SaveSongToLibrary(trackData) {
 	return new Promise(async (resolve, reject) => {
 		try {
+			if (trackData.id == null) { throw new Error("You must provide a valid library uuid") }
+			if (trackData.user_id == null) { throw new Error("You must provide a valid user id") }
+
 			const columns = Object.keys(trackData).join(", ")
 			const placeholders = Object.keys(trackData).map(() => "?").join(", ")
 			const values = Object.values(trackData)
@@ -20,27 +23,40 @@ export function SaveSongToLibrary(trackData) {
 			resolve("success")
 		} catch (e) {
 			console.log("Error saving song to library - " + e.toString())
-			reject(e.toString())
+			reject(e)
 		}
 	})
 }
 
-/** @param {String} libraryUuid */
-export function DeleteSongFromLibrary(libraryUuid) {
+/** 
+ * @param {String[]} libraryUuidArray
+ * @param {String} userId
+ **/
+export function DeleteSongFromLibrary(libraryUuidArray, userId) {
 	return new Promise(async (resolve, reject) => {
 		try {
+			if (userId == null) { throw new Error("You must provide a valid user id") }
+			if (Array.isArray(libraryUuidArray) === false || libraryUuidArray.length === 0) { throw new Error("You must provide a valid library uuid array") }
+			libraryUuidArray = libraryUuidArray.filter((uuid) => uuid != null && uuid.trim() !== "")
+			const placeholders = libraryUuidArray.map(() => "?").join(", ")
 			// Remove the song from the database
-			const query = "DELETE FROM songs WHERE id = ?"
-			const params = libraryUuid
-			Database.writeQuery(query, params)
+			const query = "DELETE FROM songs WHERE user_id = ? AND id IN (" + placeholders + ")"
+			const params = [userId, ...libraryUuidArray]
+			const result = await Database.writeQuery(query, params)
 
-			// Delete the song file from the file system
-			const filePath = path.join("music-library", libraryUuid + ".mp3")
-			// Check if the file exists
-			await access(filePath, constants.F_OK)
-			// Delete the file
-			await unlink(filePath)
-			resolve()
+			for (const libraryUuid of libraryUuidArray) {
+				try {
+					// Delete the song file from the file system
+					let filePath = path.join("music-library", libraryUuid + ".mp3")
+					// Check if the file exists
+					await access(filePath, constants.F_OK)
+					// Delete the file
+					await unlink(filePath)
+				} catch (e) {
+					console.error("Error deleting song from filesystem - " + e.toString())
+				}
+			}
+			resolve(result.changes)
 		} catch (e) {
 			reject(e)
 		}
@@ -52,9 +68,11 @@ const router = express.Router()
 
 router.get("/", async (req, res) => {
 	try {
-
-		const query = "SELECT * FROM songs"
-		const data = await Database.readQuery(query)
+		const userId = req.user?.id
+		if (userId == null) { throw new Error("A valid user id is required") }
+		const query = "SELECT * FROM songs WHERE user_id = ?"
+		const params = [userId]
+		const data = await Database.readQuery(query, params)
 		res.json(data)
 
 	} catch (error) {
@@ -64,12 +82,13 @@ router.get("/", async (req, res) => {
 
 router.delete("/delete-song", async (req, res) => {
 	try {
-		const libraryUuid = req.body?.libraryUuid
-		if (libraryUuid == null) { throw new Error("You must provide a library uuid") }
-		await DeleteSongFromLibrary(libraryUuid)
+		const idArray = req.body?.idArray
+		const userId = req.user?.id
+		let deletionCount = 0
+		deletionCount = await DeleteSongFromLibrary(idArray, userId)
 		res.send({
 			status: "success",
-			libraryUuid: libraryUuid
+			deletionCount: deletionCount
 		})
 	} catch (e) {
 		res.status(404).send({ error: e.toString() })
@@ -78,6 +97,9 @@ router.delete("/delete-song", async (req, res) => {
 
 router.post("/edit-song-data", async (req, res) => {
 	try {
+		const libraryUuid = req.body?.libraryUuid
+		const userId = req.user?.id
+
 
 		res.send()
 	} catch (e) {
